backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pymongo[srv]>=4.6.0
dnspython>=2.4.0
PyJWT>=2.8.0
Werkzeug>=3.0.0
bson>=0.5.10
cloudinary>=1.41.0
